Fixed bugs - Hover to see.
We've now fully implemented who the player was banned by, so for example: Banned by $sender->getName() which is the sender who banned the person, instead of Staff. That way, it's more professional for servers. :)
Fixed $sender->getName() from appearing to be broken in some commands.
Should now fix Command errors when attempted to execute the commands. There's one which isn't fixed, which is /tmute. We're working on fixing this soon.
Now added Ban appeals to when you get banned. Don't worry though, we plan on adding a config to change the Ban appeal link, so it's more original to you. :)
We've listened! And we've finally removed the IP address $ip, as people can potentially DDOSS / use the ip to their advantages. The $ip variable is the IP address of your wifi. We removed this because people can easily hack the ip address. I have no idea why this was added in the first place. ;-;
Now added who unbanned the user by using $sender->getName(). This also goes to unbanning an ip address.
We've also added in who the ip belongs to when they got banned. It's currently experimental, not sure if this feature will become a thing, but we'll see.
Many bug fixes, typos fixed, and grammar fixes within the code.
Fixed the $sender from being there. It isn't needed, as that doesn't go to that specific sender who banned / unbanned the user. It is still there in some cases, but we'll be sure to fix up all the codes very soon.
We've also fixed many bugs and improvements upon this.

TODO LIST
As of our todo list, we'll be adding / fixing:
* Color code fixes, make it more professional like.
* From most commands, we'll be fixing some useless codes + Adding $sender->getName() temp banned player->getName() instead of the other way round. Could be easier to see, as currently, the code is messed up. We'll be fixing soon, don't worry. :)
Fix /tmute
Fix /tblock
Add a config file for Ban Appeal links. :D

If you experience any more issues, please open an issue and report it there. Thank you.